Extract caretaker roles test fixture
