+++ App/views/CaretakerRolesMenu/CaretakerRolesMenu.spec.js
@@ -1,40 +1,20 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
 import CaretakerRolesMenu from './CaretakerRolesMenu';
 import UserGateway, { Permissions } from '../../data/UserGateway';
-import CaretakerRolesGateway from '../../data/CaretakerRolesGateway';
+import CaretakerRolesGateway from '../../data/CaretakerRoles/CaretakerRoles.gateway';
+import CaretakerRolesFixture from '../../data/CaretakerRoles/CaretakerRoles.fixture';
 import { promiseComponent, promiseValue } from '../../utils/testUtils';
 
 describe('<CaretakerRolesMenu />', () => {
   let getCurrentPermissions;
   let caretakerRolesMenu;
-  const roles = [{
-    id: 1,
-    name: 'Driver',
-    description: 'Gives rides to things'
-  }, {
-    id: 2,
-    name: 'Coordinator',
-    description: 'Helps coordinate people sign ups'
-  }, {
-    id: 3,
-    name: 'Groceries',
-    description: 'Picks up groceries'
-  }, {
-    id: 4,
-    name: 'Active Friend',
-    description: 'Gets focus out and active (eg, walks) during vulnerable times'
-  }, {
-    id: 5,
-    name: 'Chef',
-    description: 'Cooks food cause yum'
-  }];
+  const roles = CaretakerRolesFixture.allRoles;
 
   beforeEach(() => {
     getCurrentPermissions = jest.spyOn(UserGateway, 'getCurrentPermissions');
     const getAllRolesPromise = promiseValue(roles);
-    jest.spyOn(CaretakerRolesGateway, 'getAll').mockReturnValue(getAllRolesPromise);
     caretakerRolesMenu = promiseComponent(<CaretakerRolesMenu />, getAllRolesPromise);
+    jest.spyOn(CaretakerRolesGateway, 'getAll').mockReturnValue(getAllRolesPromise);
     jest.spyOn(UserGateway, 'getRoles').mockReturnValue(['driver']);
   });
